+++ pkgs/dart_mcp/example/workflow_client.dart
@@ -12,6 +12,14 @@ import 'package:cli_util/cli_logging.dart';
 import 'package:dart_mcp/client.dart';
 import 'package:google_generative_ai/google_generative_ai.dart' as gemini;
 
+/// The list of Gemini models that are accepted as a "--model" argument.
+/// Defaults to the first one in the list.
+const List<String> allowedGeminiModels = [
+  'gemini-2.5-pro-exp-03-25',
+  'gemini-2.0-flash',
+  'gemini-2.5-flash-preview-04-17',
+];
+
 void main(List<String> args) {
   final geminiApiKey = Platform.environment['GEMINI_API_KEY'];
   if (geminiApiKey == null) {
@@ -23,8 +31,13 @@ void main(List<String> args) {
   }
 
   final parsedArgs = argParser.parse(args);
+  if (parsedArgs.wasParsed('help')) {
+    print(argParser.usage);
+    exit(0);
+  }
   final serverCommands = parsedArgs['server'] as List<String>;
   final logger = Logger.standard();
+  final logFilePath = parsedArgs.option('log');
   runZonedGuarded(
     () {
       WorkflowClient(
@@ -32,12 +45,14 @@ void main(List<String> args) {
         geminiApiKey: geminiApiKey,
         verbose: parsedArgs.flag('verbose'),
         dtdUri: parsedArgs.option('dtd'),
+        model: parsedArgs.option('model')!,
         persona: parsedArgs.flag('dash') ? _dashPersona : null,
         logger: logger,
+        logFile: logFilePath != null ? File(logFilePath) : null,
       );
     },
-    (e, s) {
-      logger.stderr('$e\n$s\n');
+    (exception, stack) {
+      logger.stderr('$exception\n$stack\n');
     },
   );
 }
@@ -54,28 +69,38 @@ final argParser =
         abbr: 'v',
         help: 'Enables verbose logging for logs from servers.',
       )
+      ..addOption(
+        'log',
+        abbr: 'l',
+        help:
+            'If specified, will create the given log file and log server '
+            'traffic and diagnostic messages.',
+      )
       ..addFlag('dash', help: 'Use the Dash mascot persona.', defaultsTo: false)
       ..addOption(
         'dtd',
         help: 'Pass the DTD URI to use for this workflow session.',
-      );
+      )
+      ..addOption(
+        'model',
+        defaultsTo: allowedGeminiModels.first,
+        allowed: allowedGeminiModels,
+        help: 'Pass the name of the model to use to run inferences.',
+      )
+      ..addFlag('help', abbr: 'h', help: 'Print the usage for this command.');
 
 final class WorkflowClient extends MCPClient with RootsSupport {
-  final Logger logger;
-  int totalInputTokens = 0;
-  int totalOutputTokens = 0;
-
   WorkflowClient(
     this.serverCommands, {
     required String geminiApiKey,
+    required String model,
+    required this.logger,
     String? dtdUri,
     this.verbose = false,
-    required this.logger,
     String? persona,
+    File? logFile,
   }) : model = gemini.GenerativeModel(
-         model: 'gemini-2.5-pro-preview-03-25',
-         // model: 'gemini-2.0-flash',
-         //  model: 'gemini-2.5-flash-preview-04-17',
+         model: model,
          apiKey: geminiApiKey,
          systemInstruction: systemInstructions(persona: persona),
        ),
@@ -85,6 +110,7 @@ final class WorkflowClient extends MCPClient with RootsSupport {
        super(
          ClientImplementation(name: 'Gemini workflow client', version: '0.1.0'),
        ) {
+    logSink = _createLogSink(logFile);
     addRoot(
       Root(
         uri: Directory.current.absolute.uri.toString(),
@@ -110,6 +136,10 @@ final class WorkflowClient extends MCPClient with RootsSupport {
     _startChat();
   }
 
+  final Logger logger;
+  Sink<String>? logSink;
+  int totalInputTokens = 0;
+  int totalOutputTokens = 0;
   final StreamQueue<String> stdinQueue;
   final List<String> serverCommands;
   final List<ServerConnection> serverConnections = [];
@@ -117,6 +147,39 @@ final class WorkflowClient extends MCPClient with RootsSupport {
   final List<gemini.Content> chatHistory = [];
   final gemini.GenerativeModel model;
   final bool verbose;
+
+  Sink<String>? _createLogSink(File? logFile) {
+    if (logFile == null) {
+      return null;
+    }
+    Sink<String>? logSink;
+    logFile.createSync(recursive: true);
+    final fileByteSink = logFile.openWrite(
+      mode: FileMode.write,
+      encoding: utf8,
+    );
+    logSink = fileByteSink.transform<String>(
+      StreamSinkTransformer.fromHandlers(
+        handleData: (String data, EventSink<List<int>> innerSink) {
+          innerSink.add(utf8.encode(data));
+          // It's a log, so we want to make sure it's always up-to-date.
+          fileByteSink.flush();
+        },
+        handleError: (
+          Object error,
+          StackTrace stackTrace,
+          EventSink<List<int>> innerSink,
+        ) {
+          innerSink.addError(error, stackTrace);
+          fileByteSink.flush();
+        },
+        handleDone: (EventSink<List<int>> innerSink) {
+          innerSink.close();
+        },
+      ),
+    );
+    return logSink;
+  }
 
   void _startChat() async {
     if (serverCommands.isNotEmpty) {
@@ -349,7 +412,11 @@ final class WorkflowClient extends MCPClient with RootsSupport {
       final parts = server.split(' ');
       try {
         serverConnections.add(
-          await connectStdioServer(parts.first, parts.skip(1).toList()),
+          await connectStdioServer(
+            parts.first,
+            parts.skip(1).toList(),
+            protocolLogSink: logSink,
+          ),
         );
       } catch (e) {
         logger.stderr('Failed to connect to server $server: $e');
@@ -370,10 +437,11 @@ final class WorkflowClient extends MCPClient with RootsSupport {
         ),
       );
       final serverName = connection.serverInfo?.name ?? 'server';
-      if (result.protocolVersion != ProtocolVersion.latestSupported) {
+      if (!result.protocolVersion!.isSupported) {
         logger.stderr(
           'Protocol version mismatch for $serverName, '
-          'expected ${ProtocolVersion.latestSupported}, got '
+          'expected a version between ${ProtocolVersion.oldestSupported} and '
+          '${ProtocolVersion.latestSupported}, but got '
           '${result.protocolVersion}. Disconnecting.',
         );
         await connection.shutdown();