+++ pkgs/dart_mcp/CHANGELOG.md
@@ -6,6 +6,9 @@
 - Update file system example to support relative paths.
 - Fix a bug in notification handling where leaving off the parameters caused a
   type exception.
+- Added `--help`, `--log`, and `--model` flags to the workflow example.
+- Fixed a bug where the examples would only connect to a server with the latest
+  protocol version.
 
 ## 0.2.0